address review : part 1
